// src/graphql/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * The `Date` scalar type represents a Date
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Date: { input: any; output: any; }
  /**
   * The `DateTime` scalar type represents a DateTime
   * value as specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  DateTime: { input: any; output: any; }
  /**
   * The `GenericScalar` scalar type represents a generic
   * GraphQL scalar value that could be:
   * String, Boolean, Int, Float, List or Object.
   */
  GenericScalar: { input: any; output: any; }
  /**
   * Allows use of a JSON String for input / output from the GraphQL schema.
   *
   * Use of this type is *not recommended* as you lose the benefits of having a defined, static
   * schema (one of the key benefits of GraphQL).
   */
  JSONString: { input: any; output: any; }
  /**
   * The `Time` scalar type represents a Time value as
   * specified by
   * [iso8601](https://en.wikipedia.org/wiki/ISO_8601).
   */
  Time: { input: any; output: any; }
  /**
   * Create scalar that ignores normal serialization/deserialization, since
   * that will be handled by the multipart request spec
   */
  Upload: { input: any; output: any; }
};
// ...
export type BrandNodeConnection = {
  __typename?: 'BrandNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<BrandNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CartNodeConnection = {
  __typename?: 'CartNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<CartNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type CityNode = Node & {
  __typename?: 'CityNode';
  alternateNames?: Maybe<Scalars['String']['output']>;
  city: ServiceContactNodeConnection;
  cityModel: DownloadLogNodeConnection;
  contacts: ContactNodeConnection;
  country: CountryNode;
  displayName: Scalars['String']['output'];
  featureCode?: Maybe<Scalars['String']['output']>;
  geonameId?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  jobApplications: JobApplicationNodeConnection;
  jobpositionSet: JobPositionNodeConnection;
  latitude?: Maybe<Scalars['Float']['output']>;
  longitude?: Maybe<Scalars['Float']['output']>;
  medicalfilmsolutionSet: MedicalFilmSolutionNodeConnection;
  name: Scalars['String']['output'];
  nameAscii: Scalars['String']['output'];
  offices: OfficeNodeConnection;
  population?: Maybe<Scalars['Int']['output']>;
  profiles: ProfileNodeConnection;
  promotedprojectSet: PromotedProjectNodeConnection;
  region?: Maybe<RegionNode>;
  samplescontact: SampleContactNodeConnection;
  searchNames: Scalars['String']['output'];
  shippingAddresses: Array<ShippingAddressNode>;
  slug: Scalars['String']['output'];
  stores: StoreNodeConnection;
  timezone?: Maybe<Scalars['String']['output']>;
  vacancies: VacantNodeConnection;
};
// ...
export type CompanyTypeNodeConnection = {
  __typename?: 'CompanyTypeNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<CompanyTypeNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ContactNodeConnection = {
  __typename?: 'ContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type CountryNode = Node & {
  __typename?: 'CountryNode';
  alternateNames?: Maybe<Scalars['String']['output']>;
  citySet: CityNodeConnection;
  code2?: Maybe<Scalars['String']['output']>;
  code3?: Maybe<Scalars['String']['output']>;
  continent?: Maybe<Scalars['String']['output']>;
  formproposalSet: FormProposalNodeConnection;
  geonameId?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  nameAscii: Scalars['String']['output'];
  offices?: Maybe<OfficeNodeConnection>;
  phone?: Maybe<Scalars['String']['output']>;
  projects: ProjectNodeConnection;
  promotedprojectSet: PromotedProjectNodeConnection;
  regionSet: RegionNodeConnection;
  slug: Scalars['String']['output'];
  tld: Scalars['String']['output'];
};
// ...
export type DocumentNodeConnection = {
  __typename?: 'DocumentNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<DocumentNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type DownloadLogNodeConnection = {
  __typename?: 'DownloadLogNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<DownloadLogNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type FaqNodeConnection = {
  __typename?: 'FAQNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<FaqNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type GlobalDataNodeConnection = {
  __typename?: 'GlobalDataNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<GlobalDataNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type InstanceNode = Node & {
  __typename?: 'InstanceNode';
  brands: BrandNodeConnection;
  cartSet: CartNodeConnection;
  cartitemSet: ItemNodeConnection;
  claimSet: GlobalDataNodeConnection;
  code?: Maybe<Scalars['String']['output']>;
  companytypes: CompanyTypeNodeConnection;
  contacts: ContactNodeConnection;
  documents: DocumentNodeConnection;
  downloadlogSet: DownloadLogNodeConnection;
  emailsubscriberSet: Array<EmailSubscriberType>;
  facebookUrl: Scalars['String']['output'];
  faqs: FaqNodeConnection;
  flatpageSet: FlatpageNodeConnection;
  homesliderSet: HomeSliderNodeConnection;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instagramUrl: Scalars['String']['output'];
  instance: ServiceContactNodeConnection;
  jobApplications: JobApplicationNodeConnection;
  linkedinUrl: Scalars['String']['output'];
  logo?: Maybe<Scalars['String']['output']>;
  markets: MarketNodeConnection;
  name: Scalars['String']['output'];
  orderitemSet: OrderItemNodeConnection;
  orders: OrderNodeConnection;
  posts: PostNodeConnection;
  products: ProductNodeConnection;
  projects: ProjectNodeConnection;
  projectsInstance: PromotedProjectNodeConnection;
  /** Please insert the emails separated for commas used in samples request in form No! shopping cart */
  sampleEmails?: Maybe<Scalars['String']['output']>;
  samplescontact: SampleContactNodeConnection;
  services: ServiceNodeConnection;
  stores: StoreNodeConnection;
  storesInstance: OfficeNodeConnection;
  twitterUrl: Scalars['String']['output'];
  userProfiles: ProfileNodeConnection;
  users: UserNodeConnection;
  vacancies: VacantNodeConnection;
};
// ...
export type ItemNodeConnection = {
  __typename?: 'ItemNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ItemNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type JobApplicationNodeConnection = {
  __typename?: 'JobApplicationNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<JobApplicationNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type JobPositionNodeConnection = {
  __typename?: 'JobPositionNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<JobPositionNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type MedicalFilmSolutionNodeConnection = {
  __typename?: 'MedicalFilmSolutionNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<MedicalFilmSolutionNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type Node = {
  /** The ID of the object. */
  id: Scalars['ID']['output'];
};

export type OdourTestNode = Node & {
  __typename?: 'OdourTestNode';
  additionalTest: AdditionalTestsNode;
  dry?: Maybe<CrockingResistanceITestDry>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  method?: Maybe<Scalars['String']['output']>;
  range?: Maybe<CrockingResistanceITestRange>;
  wet?: Maybe<CrockingResistanceITestWet>;
};

export type OfficeNode = Node & {
  __typename?: 'OfficeNode';
  address: Scalars['String']['output'];
  /** List of all emails (primary and secondary). */
  allEmails?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  business?: Maybe<Scalars['String']['output']>;
  city: CityNode;
  email: Scalars['String']['output'];
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  instance?: Maybe<InstanceNode>;
  latitude: Scalars['Float']['output'];
  longitude: Scalars['Float']['output'];
  name: Scalars['String']['output'];
  phone: Scalars['String']['output'];
};

export type OfficeNodeConnection = {
  __typename?: 'OfficeNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<OfficeNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ProfileNodeConnection = {
  __typename?: 'ProfileNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ProfileNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type PromotedProjectNodeConnection = {
  __typename?: 'PromotedProjectNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<PromotedProjectNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type RegionNode = Node & {
  __typename?: 'RegionNode';
  alternateNames?: Maybe<Scalars['String']['output']>;
  citySet: CityNodeConnection;
  country: CountryNode;
  displayName: Scalars['String']['output'];
  geonameCode?: Maybe<Scalars['String']['output']>;
  geonameId?: Maybe<Scalars['Int']['output']>;
  /** The ID of the object. */
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  nameAscii: Scalars['String']['output'];
  slug: Scalars['String']['output'];
};
// ...
export type SampleContactNodeConnection = {
  __typename?: 'SampleContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<SampleContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ServiceContactNodeConnection = {
  __typename?: 'ServiceContactNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<ServiceContactNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};
// ...
export type ShippingAddressNode = {
  __typename?: 'ShippingAddressNode';
  address: Scalars['String']['output'];
  city?: Maybe<CityNode>;
  cityName?: Maybe<Scalars['String']['output']>;
  company: Scalars['String']['output'];
  firstName: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  lastName: Scalars['String']['output'];
  /** Indicates whether this is the main/default shipping address. */
  mainAddress: Scalars['Boolean']['output'];
  order: OrderNode;
  source: Scalars['String']['output'];
  zipCode?: Maybe<Scalars['Int']['output']>;
};
// ...
export type StoreNodeConnection = {
  __typename?: 'StoreNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<StoreNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type VacantNodeConnection = {
  __typename?: 'VacantNodeConnection';
  /** Contains the nodes in this connection. */
  edges: Array<Maybe<VacantNodeEdge>>;
  /** Pagination data for this connection. */
  pageInfo: PageInfo;
};